# Prefer strings.Builder over unsafebytes