--- a/bottom/internal/unsafebytes/unsafebytes.go
+++ b/bottom/internal/unsafebytes/unsafebytes.go
@@ -1,9 +1,0 @@
-package unsafebytes
-
-import "unsafe"
-
-// String converts a byte slice to a string in an unsafe, non-copy fashion. It
-// works similarly to strings.Builder's String method.
-func String(bytes []byte) string {
-	return *(*string)(unsafe.Pointer(&bytes))
-}