This pull request removes the use of `unsafe` and instead prefers `strings.Builder`. This causes a slight performance loss during encoding and none during decoding:

```
# Old: unsafe
# New: strings.Builder

name          old time/op    new time/op    delta
Encode-8         276ns ±18%     292ns ± 3%  +5.61%  (p=0.010 n=15+15)
EncodeTo-8      1.87µs ± 3%    1.85µs ± 7%    ~     (p=0.056 n=13+15)
Decode-8        3.16µs ± 2%    3.18µs ± 7%    ~     (p=0.784 n=13+14)
DecodeFrom-8    3.74µs ± 4%    3.73µs ± 2%    ~     (p=0.562 n=15+13)

name          old speed      new speed      delta
Encode-8      43.0MB/s ±12%  41.1MB/s ± 3%  -4.38%  (p=0.018 n=14+15)
EncodeTo-8     175MB/s ± 3%   177MB/s ± 6%    ~     (p=0.060 n=13+15)
Decode-8       103MB/s ± 2%   103MB/s ± 6%    ~     (p=0.756 n=13+14)
DecodeFrom-8  87.4MB/s ± 4%  87.7MB/s ± 2%    ~     (p=0.563 n=15+13)

```