@nihaals 2021-02-23
What's the advantage of avoiding unsafe here? Is there potential UB?