@diamondburned 2021-02-23
No, it's just cleaner. There's just no unsafe.