@nihaals 2021-02-23
I'm unsure, I think performance is an interesting goal, especially with such a small project. Code maintainability isn't a big issue as it would require a very odd implementation for the code to become unreadable or too spaghetti. The only required change I see in the future is #1, other than that, there shouldn't need to be any large changes to functionality or features, so I'm not sure how much of a positive difference this will have (compared to even a fairly small increase of performance).