@diamondburned 2021-02-23
The performance loss was likely because of the preoccupied load on my laptop. A higher sample size would make up for the offset.