@nihaals 2021-02-24
Based on the GitHub Actions log ([new](https://github.com/bottom-software-foundation/bottom-go/runs/1956248790#step:5:1), [previous](https://github.com/bottom-software-foundation/bottom-go/runs/1893592238#step:5:1)):

## Previous
```
BenchmarkEncode-2       	 7240875	       167 ns/op	  71.80 MB/s
BenchmarkEncodeTo-2     	  805804	      1317 ns/op	 248.26 MB/s
BenchmarkDecode-2       	  509618	      2067 ns/op	 158.24 MB/s
BenchmarkDecodeFrom-2   	  427752	      2345 ns/op	 139.46 MB/s
```

## New
```
BenchmarkEncode-2       	 6624477	       174 ns/op	  68.83 MB/s
BenchmarkEncodeTo-2     	  915034	      1251 ns/op	 261.34 MB/s
BenchmarkDecode-2       	  607250	      2019 ns/op	 161.98 MB/s
BenchmarkDecodeFrom-2   	  564886	      2089 ns/op	 156.51 MB/s
```

This looks like random variation so I agree, doesn't seem to be a real impact on performance